+++ CHANGELOG.md
@@ -12,6 +12,7 @@
 
 - **ArchLinux API** Prevent errors when response or results is None/empty
 - **Logger** Change logger level back to WARNING
+- **Package Handler** Harden extract_upstream_url_nvchecker against missing package keys
 
 # 1.0.9 (2025-07-24)